0.0.5a Added solution.
