+++ src/main/webapp/web.xml
@@ -92,14 +92,6 @@
         <url-pattern>/removeProductFromShopCart</url-pattern>
     </servlet-mapping>
     <servlet>
-        <servlet-name>CreateShopCart</servlet-name>
-        <servlet-class>internet.shop.controllers.shoppingcart.CreateShopCartController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>CreateShopCart</servlet-name>
-        <url-pattern>/CreateShopCart</url-pattern>
-    </servlet-mapping>
-    <servlet>
         <servlet-name>CreateOrder</servlet-name>
         <servlet-class>internet.shop.controllers.order.CreateOrderController</servlet-class>
     </servlet>
@@ -132,19 +124,11 @@
         <url-pattern>/orders/all</url-pattern>
     </servlet-mapping>
     <servlet>
-        <servlet-name>generateUsers</servlet-name>
-        <servlet-class>internet.shop.controllers.generators.GenerateUsersController</servlet-class>
+        <servlet-name>InjectData</servlet-name>
+        <servlet-class>internet.shop.controllers.generators.InjectDataController</servlet-class>
     </servlet>
     <servlet-mapping>
-        <servlet-name>generateUsers</servlet-name>
-        <url-pattern>/generateUsers</url-pattern>
-    </servlet-mapping>
-    <servlet>
-        <servlet-name>generateProducts</servlet-name>
-        <servlet-class>internet.shop.controllers.generators.GenerateProductsController</servlet-class>
-    </servlet>
-    <servlet-mapping>
-        <servlet-name>generateProducts</servlet-name>
-        <url-pattern>/generateProducts</url-pattern>
+        <servlet-name>InjectData</servlet-name>
+        <url-pattern>/InjectData</url-pattern>
     </servlet-mapping>
 </web-app>
